+++ audio_be/src/commonMain/kotlin/voices/Voice.kt
@@ -23,7 +23,7 @@ sealed interface Voice {
         /** Number of frames to process per block */
         val blockFrames: Int,
         /** Shared mono buffer for voice rendering */
-        val voiceBuffer: DoubleArray,
+        val voiceBuffer: FloatArray,
         /** Shared buffer for frequency/pitch modulation calculations */
         val freqModBuffer: DoubleArray,
         /** Shared scratch buffer pool for SignalGen composition operators */
